scripts/requirements.txt
# Core dependencies
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0
google-auth>=2.0.0
pytz>=2023.3

# HTTP and networking
requests>=2.31.0
urllib3>=2.0.0
idna>=3.0
certifi>=2023.0.0

# Text processing
python-slugify>=8.0.0
charset-normalizer>=3.0.0

# Utilities
six>=1.16.0
cachetools>=5.0.0
pyasn1>=0.5.0
pyasn1-modules>=0.3.0
rsa>=4.9
